Rename package to 'contentful_cli'
Problem: The existing package name 'contentful' clashes with the
official Contentful Python delivery API library.

Solution: Rename to contentful_cli. Major version increment since this
would break any code attempting to use this as a library.
